Update the script to restore db from specific date first

+++ gigadb/app/tools/files-url-updater/reset_database.sh
@@ -49,40 +49,18 @@ docker-compose down -v
 # Terminate other connections to RDS instance
 export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c  "SELECT pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname='gigadb';"
 
-# Load the latest dump in RDS using native postgresql client
-if [[ $downloadLatestRestoreStatus -eq 0 && $convertStatus -eq 0 ]];then
-  echo "Loading gigadbv3_${latest}_v${version}.backup"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "drop database if exists $DB_PG_DATABASE"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "create database $DB_PG_DATABASE owner $DB_PG_USER"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -h $DB_PG_HOST -p 5432 < converted/gigadbv3_${latest}_v${version}.backup
-  loadStatus=$?
-elif [[ -f converted/gigadbv3_${twoDaysAgo}_v${version}.backup ]];then
-  echo "Loading gigadbv3_${twoDaysAgo}_v${version}.backup from two days ago"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "drop database if exists $DB_PG_DATABASE"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "create database $DB_PG_DATABASE owner $DB_PG_USER"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -h $DB_PG_HOST -p 5432 < converted/gigadbv3_${twoDaysAgo}_v${version}.backup
-  loadStatus=$?
-elif [[ -f converted/gigadbv3_${threeDaysAgo}_v${version}.backup ]];then
-  echo "Loading gigadbv3_${threeDaysAgo}_v${version}.backup from three days ago"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "drop database if exists $DB_PG_DATABASE"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "create database $DB_PG_DATABASE owner $DB_PG_USER"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -h $DB_PG_HOST -p 5432 < converted/gigadbv3_${threeDaysAgo}_v${version}.backup
-  loadStatus=$?
-else
-  echo "Loading default backup /home/centos/database_bootstrap.backup"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "drop database if exists $DB_PG_DATABASE"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "create database $DB_PG_DATABASE owner $DB_PG_USER"
-  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -h $DB_PG_HOST -p 5432 < "$defaultDB"
-  loadStatus=$?
-fi
-exit $loadStatus
-
-# Load the specific dump in RDS using native postgresql client
-if [[ $downloadSpecificRestoreStatus -eq 0 && $convertStatus -eq 0 ]];then
+# Load the backup dump to RDS using native postgresql client
+if [[ $downloadSpecificRestoreStatus -eq 0 && $convertStatus -eq 0 && -z $downloadLatestRestoreStatus ]];then
   echo "Loading gigadbv3_${backupDate}_v${version}.backup"
   export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "drop database if exists $DB_PG_DATABASE"
   export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "create database $DB_PG_DATABASE owner $DB_PG_USER"
   export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -h $DB_PG_HOST -p 5432 < converted/gigadbv3_${backupDate}_v${version}.backup
+  loadStatus=$?
+elif [[ $downloadLatestRestoreStatus -eq 0 && $convertStatus -eq 0 && -z $downloadSpecificRestoreStatus ]];then
+  echo "Loading gigadbv3_${latest}_v${version}.backup"
+  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "drop database if exists $DB_PG_DATABASE"
+  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -d postgres -h $DB_PG_HOST -p 5432 -c "create database $DB_PG_DATABASE owner $DB_PG_USER"
+  export PGPASSWORD=$DB_PG_PASSWORD; psql -U $DB_PG_USER -h $DB_PG_HOST -p 5432 < converted/gigadbv3_${latest}_v${version}.backup
   loadStatus=$?
 elif [[ -f converted/gigadbv3_${twoDaysAgo}_v${version}.backup ]];then
   echo "Loading gigadbv3_${twoDaysAgo}_v${version}.backup from two days ago"
